// packages/back-end/types/query.ts
import { QueryLanguage } from "./datasource";

export type QueryStatus = "running" | "failed" | "succeeded";

export type QueryPointer = {
  query: string;
  status: QueryStatus;
  name: string;
};
export type Queries = QueryPointer[];

export interface QueryInterface {
  id: string;
  organization: string;
  datasource: string;
  language: QueryLanguage;
  query: string;
  status: QueryStatus;
  createdAt: Date;
  startedAt: Date;
  finishedAt: Date;
  heartbeat: Date;
  // eslint-disable-next-line
  result?: Record<string, any>;
  error?: string;
}